+++ src/lib/pages/instantiate/component/InstantiateOffchainForm.tsx
@@ -90,6 +90,7 @@ export const InstantiateOffChainForm = observer(
         )}
         <OffChainForm<OffchainDetail>
           state={offchainState}
+          label={contractLabel}
           control={control}
           setTagsValue={setTagsValue}
           setContractListsValue={setContractListsValue}